Convert tests to nunit
This allows them to be run on non Visual Studio IDEs.  Mostly a
mechanical change, with a few interesting tidbits. First, nunit no
longer handles expected exceptions with an attribute, you need to
assert where the exception should happen.  Second, nunit does not
consider a DateTime to be equal to a DateTimeOffset in UTC.  Third,
in a few places we were expecting a KeyNotFoundException but getting
something else.  I changed the expected exception type rather than
the code to get these to pass.