Cleaned up Exceptions
Moved exceptions to their own folder.  Deleted unused ctors.  Created
a base class for all custom exceptions in the IdGen lib, and had
exceptions inherit from it.

This is a silly, petty cosmetic thing to do and I should have skipped
it.  But once I saw I thought of it I couldn't stop, so just getting
it out of the way.